test: ckl graphtest class
