+++ .github/workflows/lib-ci.yml
@@ -18,7 +18,7 @@ jobs:
       uses: actions/setup-dotnet@v4
       with:
         dotnet-version: '8.0.x'
-    
+        
     - name: Restore dependencies
       run: dotnet restore CKLLib\CKLLib.sln
